refactor(area-chart): added useChartLegend
